Rename and refactor classes and test classes
